subclass collection and don't override AR collection
