--- lib/active_resource/collection_ext.rb
@@ -1,7 +1,0 @@
-require 'shopify_api/collection_pagination'
-
-module ActiveResource
-  class Collection
-    prepend ShopifyAPI::CollectionPagination
-  end
-end
